### 2025-07-04 #8542 @editorialbot
[PRE REVIEW]: PyMilo: A Python Library for ML I/O
**Submitting author:** @ahreccese (<a href="http://orcid.org/0009-0000-0638-2263">AmirHosein Rostami</a>)
**Repository:** https://github.com/openscilab/pymilo
**Branch with paper.md** (empty if default branch): dev
**Version:** v1.3
