### 2025-07-10 #8542 @crvernon
:wave: @shrishar - when a topic editor is assigned they will request reviewers.  It is noted that you are available in this thread.  You can add your name to the reviewer list here: https://reviewers.joss.theoj.org/join for consideration in other reviews as well.
