### 2025-08-06 00b5664
link `CONTRIBUTING.md` to `README.md` (#223)
