### 2025-08-06 #8542 @AHReccese
@abhishektiwari
The link to the contribution guideline is now prominently added to the [README.md](https://github.com/openscilab/pymilo/tree/dev?tab=readme-ov-file#contributing) ✅
