+++ paper/paper.bib
@@ -41,15 +41,17 @@
   doi       = {10.1148/ryai.210127},
 }
 
-@article{bodimani2024assessing,
-  title={Assessing The Impact of Transparent AI Systems in Enhancing User Trust and Privacy},
-  author={Bodimani, Meghasai},
-  journal={Journal of Science \& Technology},
-  volume={5},
-  number={1},
-  pages={50--67},
-  year={2024},
-  doi={10.55662/JST.2024.5102}
+@article{bodimani2024assessing, 
+place={Ahmedabad, India},
+title={Assessing The Impact of Transparent AI Systems in Enhancing User Trust and Privacy}, 
+volume={5}, 
+url={https://thesciencebrigade.com/jst/article/view/68}, 
+number={1}, 
+journal={Journal of Science &amp; Technology}, 
+author={Bodimani, Meghasai}, 
+year={2024}, 
+month={Feb.}, 
+pages={50–67} 
 }
 
 @misc{Brownlee2018,
