+++ paper/paper.bib
@@ -42,16 +42,15 @@
 }
 
 @article{bodimani2024assessing, 
-place={Ahmedabad, India},
-title={Assessing The Impact of Transparent AI Systems in Enhancing User Trust and Privacy}, 
-volume={5}, 
-url={https://thesciencebrigade.com/jst/article/view/68}, 
-number={1}, 
-journal={Journal of Science &amp; Technology}, 
-author={Bodimani, Meghasai}, 
-year={2024}, 
-month={Feb.}, 
-pages={50–67} 
+  title={Assessing The Impact of Transparent AI Systems in Enhancing User Trust and Privacy}, 
+  volume={5}, 
+  url={https://thesciencebrigade.com/jst/article/view/68}, 
+  number={1}, 
+  journal={Journal of Science &amp; Technology}, 
+  author={Bodimani, Meghasai}, 
+  year={2024}, 
+  month={Feb.}, 
+  pages={50–67}, 
 }
 
 @misc{Brownlee2018,
