+++ CHANGELOG.md
@@ -6,7 +6,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- JOSS paper
 ### Changed
 - `README.md` updated
 - `__init__` function in `PyMiloServer`
