### 2025-08-08 77c36e0
Update/paper (#224)
* citation updated

* `CHANGELOG.md` updated

* apply feedback
