### 2025-08-08 #8542 @AHReccese
Yes, sure. It is merged. @abhishektiwari

The previous DOI was extracted from the paper itself. I have attached the associated paper, and the DOI is included in it. However, since it did not resolve properly, I removed the DOI from the associated reference.

[Assessing+The+Impact+of+Transparent+Ai+Systems+in+Enhancing+User+Trust+and+Privacy.pdf](https://github.com/user-attachments/files/21689705/Assessing%2BThe%2BImpact%2Bof%2BTransparent%2BAi%2BSystems%2Bin%2BEnhancing%2BUser%2BTrust%2Band%2BPrivacy.pdf)

I will share my suggestions for potential reviewers here as soon as possible.
