### 2025-08-15 11331a8
Integration of client & ml model id in end-to-end ml streaming flow (#196)
* `/register/` endpoint to register a new client

* `/request_model_id/` endpoint to request a new model id

* `/clients/` endpoint to get all clients

* `/client/models/` endpoint to request a specific model within a specific client

* integration with previous 4 endpoints in `RESTClientCommunicator`

* update previous endpoints to apply client/model id check/filter

* pass download information to `PyMiloServer` in `WebSocketServerCommunicator`

* `uuid` used to generate universal unique ids

* use `HTTPException` to h
[…]
