+++ tests/test_ml_streaming/scenarios/scenario1.py
@@ -28,16 +28,20 @@ def scenario1(compression_method, communication_protocol):
         communication_protocol=CommunicationProtocol[communication_protocol],
         )
 
-    # 3.
+    # 3. get client id + get ml model id [from remote server]
+    client.register()
+    client.register_ml_model()
+
+    # 4.
     result = client.predict(x_test)
     mse_before = mean_squared_error(y_test, result)
 
-    # 4.
-    client.upload()
     # 5.
+    client.upload()
+    # 6.
     client.download()
 
-    # 6.
+    # 7.
     result = client.predict(x_test)
     mse_after = mean_squared_error(y_test, result)
     return np.abs(mse_after-mse_before)
