### 2025-08-24 #8542 @abhishektiwari
@diningphil @isabelizimm @ag-chirag, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "PyMilo: A Python Library for ML I/O".

Based on your experience and research interest, we think you might be able to provide a great review of this submission.

Thanks
Abhi
