### 2025-08-24 #8542 @abhishektiwari
@ajinkyakadam @rakeshcusat - 👋👋👋 - Not sure if you saw my last message but will you be open to review this JOSS submission.

> [@ajinkyakadam](https://github.com/ajinkyakadam) [@rakeshcusat](https://github.com/rakeshcusat), would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "PyMilo: A Python Library for ML I/O".
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
>
> Based on your experience and research interest, we think you might be able to provide a great review of this submission.
>
> Thanks Abhi
