### 2025-10-07 e983359
mention `safetensors`
