### 2025-10-07 6c96ae6
explain `transparency` and `end-to-end fidelity`
