### 2025-10-07 e9f8b56
update figure to clarify framework support
