### 2025-10-07 c4cf175
update figure caption
