### 2025-10-09 eafd60a
minor wording update
