### 2025-10-10 f3215b1
Update/paper feedback (#234)
* add `safe tensors` ref

* mention `safetensors`

* explain `transparency` and `end-to-end fidelity`

* update figure to clarify framework support

* update figure caption

* minor wording update
