### 2025-11-04 89529f9
drop the `Type` column
