### 2025-11-04 d1ced89
update `enum` field mentioning, explain in text.
