### 2025-11-04 7acd226
Add/documentation (#237)
* add API reference for Export, Import, PymiloServer, PymiloClient to `README.md`

* gitignore the pdf of paper refs

* drop the `Type` column

* update `enum` field mentioning, explain in text.
