### 2025-11-04 e30d686
fix : Python 3.14 added to setup.py
