### 2025-11-04 f487626
doc : CHANGELOG.md updated
