### 2025-11-04 548ab07
Python 3.14 (#242)
* fix : Python 3.14 added to setup.py

* fix : Python 3.14 added to test.yml

* doc : CHANGELOG.md updated
