### 2025-11-10 #8858 @abhishektiwari
Hi @AHReccese - Re issue-5 raised by @idoby. As of now security guarantees discussed in the paper is not fully supported, we will need to update the paper to call that out. Overall comparison table 1 lacks supporting details/references, objectivity, hence I strongly recommend to remove this table from paper. Such comparison tables need to be backed by details typically covered by a review paper and not a software paper.
