### 2025-11-18 3f1cc80
Refine security explanation in PyMilo documentation
