+++ paper/paper.md
@@ -81,7 +81,7 @@ The machine learning community still lacks a safe and transparent end-to-end mod
 
 PyMilo is proposed to address the above gaps. It is an open-source Python library that provides an end-to-end solution for exporting and importing machine learning models in a safe, non-executable, and human-readable format such as JSON. PyMilo serializes trained models into a transparent format and fully reconstructs them without structural changes, preserving their original functionality and behavior. 
 This process does not affect inference time or performance and imports models on any target device without additional dependencies, enabling seamless execution in inference mode. 
-While PyMilo may import functions from widely used scientific libraries during deserialization to restore model behavior (for example NumPy or SciPy), the JSON representation itself never contains executable code; any residual risk is therefore inherited from those already trusted dependencies rather than introduced by the serialization format. Future releases will further restrict this mechanism through a clearly defined whitelist of importable functions and, ultimately, a dedicated transparent function serializer that removes the need for arbitrary imports.
+While PyMilo may import functions from widely used scientific libraries during deserialization to restore model behavior (for example, NumPy or SciPy), the JSON representation itself never contains executable code; any remaining security risk is therefore inherited from these already-trusted dependencies rather than introduced by PyMilo’s serialization mechanism.
 PyMilo benefits a wide range of stakeholders, including machine learning engineers, data scientists, and AI practitioners, by facilitating the development of more transparent and accountable AI systems. Furthermore, researchers working on transparent AI [@rauker2023toward], user privacy in ML [@bodimani2024assessing], and safe AI [@macrae2019governing] can use PyMilo as a framework that provides transparency and safety in the machine learning environment.
 
 # References
