### 2025-11-19 c771664
apply editor feedback (#245)
* apply editor feedback

* Refine security explanation in PyMilo documentation
