### 2025-11-27 cabdb64
Update README.md with version compatibility note for ML streaming feature
