### 2025-11-27 647cbec
Update version to 1.4 in documentation and code files
