### 2025-11-27 9840ee8
Update CHANGELOG.md to reflect version 1.4 changes, including added and changed sections, and update links for version comparison.
