### 2025-11-28 8ee657a
Remove empty sections from CHANGELOG.md for unreleased changes.
