### 2025-11-28 de159bc
Update README.md to clarify version compatibility for ML streaming over WebSocket, removing redundant information.
