### 2025-11-28 bfdf3c1
Update CHANGELOG.md to add an unreleased section for future changes.
