### 2025-11-28 894f1f9
Update README.md to clarify the stability status of the ML streaming feature, indicating it is under construction and not yet stable.
