### 2025-11-29 451bc7d
Update CHANGELOG.md to reflect the final changes for version 1.4, including numerous new functions in PymiloClient, RESTClientCommunicator, and PymiloServer, as well as modifications to existing functions across these components.
