### 2025-11-30 b76b69b
Version 1.4 (#247)
* Update README.md with version compatibility note for ML streaming feature

* Update version to 1.4 in documentation and code files

* Update CHANGELOG.md to reflect version 1.4 changes, including added and changed sections, and update links for version comparison.

* Remove empty sections from CHANGELOG.md for unreleased changes.

* Update README.md to clarify version compatibility for ML streaming over WebSocket, removing redundant information.

* Update CHANGELOG.md to add an unreleased section for future changes.

* Update README.md to clarify the stability status of th
[…]
