### 2025-12-01 51ca42d
Update release date for version 1.4 in CHANGELOG
