+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [1.4] - 2025-11-30
+## [1.4] - 2025-12-01
 ### Added
 - `get_allowed_models` function in `PymiloClient`
 - `get_allowance` function in `PymiloClient`
