### 2025-12-14 #8858 @abhishektiwari
@AHReccese

Please see following corrections. Once done let me know, I will review one more time before making recommendation. We nearly there.

Line: "SafeTensors" should be "Safetensors"
Line 23: "Currently (v1.3), PyMilo" should "Currently, PyMilo  (v1.4)"
Line 29: "high-stake domains" should be "high-stakes domains"

Subject-verb disagreements:

Lines 41-42: "While these methods preserve full model fidelity within the Python ecosystem, it poses serious security risks" should be "While these methods preserve full model fidelity within the Python ecosystem, they pose serious security risks and lack transparency"

Line 60: "This reflects predictive performance degradation and highlight the risks of silent behavioral drift ..." should be "This reflects predictive performance degradation and highlights the risks of silent behavioral drift ..."

Reference formatting issues:

Line 109 (References): "Journal of Science &Amp; Technology" appears to have an improperly rendered HTML entity. Should be "Journal of Science & Technology"

**Reference title should match original publication/source**. Most of references titles have incorrect capitalisation or incorrect formatting relative to original title. Please check all references for correctness. Based on my review, following are correct titles :

- ONNX (Open Neural Network Exchange)
- Assessing The Impact of Transparent AI Systems in Enhancing User Trust and Privacy
- Save and Load Machine Learning Models in Python with scikit-learn
- On Extending the Automatic Test Markup Language (ATML) for Machine Learning
- Reusing Deep Learning Models: Challenges and Directions in Software Engineering
- Assessing Methods and Tools to Improve Reporting, Increase Transparency, and Reduce Failures in Machine Learning Applications in Health Care
- PMML: An Open Standard for Sharing Models not (_PMML: An Open Standard for Sharing Models_)
- Safetensors - ML Safer for All
- Analysis of Failures and Risks in Deep Learning Model Converters: A Case Study in the ONNX Ecosystem
- Tensorflow.js vs Tensorflow (Python)
- How Do Model Export Formats Impact the Development of ML-Enabled Systems? A Case Study on Model Integration
- pickle — Python object serialization
- Towards Understanding the Faults of JavaScript-Based Deep Learning Systems
- Machine Learning in Python: Main Developments and Technology Trends in Data Science, Machine Learning, and Artificial Intelligence
- Toward Transparent AI: A Survey on Interpreting the Inner Structures of Deep Neural Networks
- TensorFlow.js: Machine Learning for the Web and Beyond
