+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{Raschka2020,
   author    = {Sebastian Raschka and Joshua Patterson and Corey Nolet},
-  title     = {Machine learning in Python: Main developments and technology trends in data science, machine learning, and artificial intelligence},
+  title     = {Machine Learning in Python: Main Developments and Technology Trends in Data Science, Machine Learning, and Artificial Intelligence},
   journal   = {Information},
   volume    = {11},
   number    = {4},
@@ -21,7 +21,7 @@
 }
 
 @inproceedings{davis2023reusing,
-  title={Reusing deep learning models: Challenges and directions in software engineering},
+  title={Reusing Deep Learning Models: Challenges and Directions in Software Engineering},
   author={Davis, James C and Jajal, Purvish and Jiang, Wenxin and Schorlemmer, Taylor R and Synovic, Nicholas and Thiruvathukal, George K},
   booktitle={2023 IEEE John Vincent Atanasoff International Symposium on Modern Computing (JVA)},
   pages={17--30},
@@ -32,7 +32,7 @@
 
 @article{Garbin2022,
   author    = {Cristina Garbin and Osvaldo Marques},
-  title     = {Assessing methods and tools to improve reporting, increase transparency, and reduce failures in machine learning applications in health care},
+  title     = {Assessing Methods and Tools to Improve Reporting, Increase Transparency, and Reduce Failures in Machine Learning Applications in Health Care},
   journal   = {Radiology: Artificial Intelligence},
   volume    = {4},
   number    = {2},
@@ -46,7 +46,7 @@
   volume={5}, 
   url={https://thesciencebrigade.com/jst/article/view/68}, 
   number={1}, 
-  journal={Journal of Science &amp; Technology}, 
+  journal={Journal of Science \& Technology}, 
   author={Bodimani, Meghasai}, 
   year={2024}, 
   month={Feb.}, 
@@ -55,7 +55,7 @@
 
 @misc{Brownlee2018,
   author    = {Jason Brownlee},
-  title     = {Save and load machine learning models in Python with scikit-learn},
+  title     = {Save and Load Machine Learning Models in Python with scikit-learn},
   howpublished = {\url{https://machinelearningmastery.com/save-load-machine-learning-models-python-scikit-learn/}},
   year      = {2018},
   note      = {Accessed: 2024-05-22}
@@ -70,21 +70,21 @@
 
 @software{onnx,
   author       = {Bai, Junjie and Lu, Fang and Zhang, Ke and others},
-  title        = {ONNX: Open Neural Network Exchange},
+  title        = {ONNX (Open Neural Network Exchange)},
   url          = {https://github.com/onnx/onnx},
   version      = {1.18.0},
   date         = {2025-05-12},
 }
 
 @article{pmml,
-  title={PMML: An open standard for sharing models},
+  title={PMML: An Open Standard for Sharing Models},
   author={Guazzelli, Alex and Zeller, Michael and Lin, Wen-Ching and Williams, Graham},
   year={2009},
   doi={10.32614/RJ-2009-010}
 }
 
 @article{jajal2023analysis,
-  title={Analysis of failures and risks in deep learning model converters: A case study in the onnx ecosystem},
+  title={Analysis of Failures and Risks in Deep Learning Model Converters: A Case Study in the ONNX Ecosystem},
   author={Jajal, Purvish and Jiang, Wenxin and Tewari, Arav and Kocinare, Erik and Woo, Joseph and Sarraf, Anusha and Lu, Yung-Hsiang and Thiruvathukal, George K and Davis, James C},
   journal={arXiv preprint arXiv:2303.17708},
[…]
