### 2025-12-14 30654e2
drop `algorithm` from init
