+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+### Changed
+- Test system modified
+### Removed
+
 ## [1.4] - 2025-12-01
 ### Added
 - `get_allowed_models` function in `PymiloClient`
