### 2025-12-14 79de756
Apply Editor's final feedback on paper (#251)
* Update paper.md to reflect version 1.4 changes, including corrections to text for clarity and accuracy, such as the update of the version number and adjustments to phrasing for improved readability.

* the title of references updated

* drop `algorithm` from init

* drop `multi_class ` from init

* Update CHANGELOG.md
