### 2025-12-15 d0fa466
Codebeat Badge (#252)
* doc : codebeat badge removed

* doc : CHANGELOG.md updated
