+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 ### Changed
 - Test system modified
+- `README.md` updated
 ### Removed
 
 ## [1.4] - 2025-12-01
