### 2025-12-17 6ba5d30
Update titles in bibliography entries to capitalize each significant word for consistency and improved readability.
