+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{Raschka2020,
   author    = {Sebastian Raschka and Joshua Patterson and Corey Nolet},
-  title     = {Machine Learning in Python: Main Developments and Technology Trends in Data Science, Machine Learning, and Artificial Intelligence},
+  title     = {{M}achine {L}earning in {P}ython: {M}ain {D}evelopments and {T}echnology {T}rends in {D}ata {S}cience, {M}achine {L}earning, and {A}rtificial {I}ntelligence},
   journal   = {Information},
   volume    = {11},
   number    = {4},
@@ -12,7 +12,7 @@
 @inproceedings{parida2025model,
   author={Parida, Shreyas Kumar and Gerostathopoulos, Ilias and Bogner, Justus},
   booktitle={2025 IEEE/ACM 4th International Conference on AI Engineering – Software Engineering for AI (CAIN)}, 
-  title={How Do Model Export Formats Impact the Development of ML-Enabled Systems? A Case Study on Model Integration}, 
+  title={{H}ow {D}o {M}odel {E}xport {F}ormats {I}mpact the {D}evelopment of {ML}-{E}nabled {S}ystems? {A} {C}ase {S}tudy on {M}odel {I}ntegration}, 
   year={2025},
   volume={},
   number={},
@@ -21,7 +21,7 @@
 }
 
 @inproceedings{davis2023reusing,
-  title={Reusing Deep Learning Models: Challenges and Directions in Software Engineering},
+  title={{R}eusing {D}eep {L}earning {M}odels: {C}hallenges and {D}irections in {S}oftware {E}ngineering},
   author={Davis, James C and Jajal, Purvish and Jiang, Wenxin and Schorlemmer, Taylor R and Synovic, Nicholas and Thiruvathukal, George K},
   booktitle={2023 IEEE John Vincent Atanasoff International Symposium on Modern Computing (JVA)},
   pages={17--30},
@@ -32,7 +32,7 @@
 
 @article{Garbin2022,
   author    = {Cristina Garbin and Osvaldo Marques},
-  title     = {Assessing Methods and Tools to Improve Reporting, Increase Transparency, and Reduce Failures in Machine Learning Applications in Health Care},
+  title     = {{A}ssessing {M}ethods and {T}ools to {I}mprove {R}eporting, {I}ncrease {T}ransparency, and {R}educe {F}ailures in {M}achine {L}earning {A}pplications in {H}ealth {C}are},
   journal   = {Radiology: Artificial Intelligence},
   volume    = {4},
   number    = {2},
@@ -42,7 +42,7 @@
 }
 
 @article{bodimani2024assessing, 
-  title={Assessing The Impact of Transparent AI Systems in Enhancing User Trust and Privacy}, 
+  title={{A}ssessing {T}he {I}mpact of {T}ransparent {AI} {S}ystems in {E}nhancing {U}ser {T}rust and {P}rivacy}, 
   volume={5}, 
   url={https://thesciencebrigade.com/jst/article/view/68}, 
   number={1}, 
@@ -55,7 +55,7 @@
 
 @misc{Brownlee2018,
   author    = {Jason Brownlee},
-  title     = {Save and Load Machine Learning Models in Python with scikit-learn},
+  title     = {{S}ave and {L}oad {M}achine {L}earning {M}odels in {P}ython with scikit-learn},
   howpublished = {\url{https://machinelearningmastery.com/save-load-machine-learning-models-python-scikit-learn/}},
   year      = {2018},
   note      = {Accessed: 2024-05-22}
@@ -63,28 +63,28 @@
 
 @misc{PythonPickleDocs,
   author       = {{Python Software Foundation}},
[…]
