### 2025-12-18 ace67df
Update paper.bib
