### 2025-12-18 b1f8cd0
Update paper/paper.bib
