+++ paper/paper.bib
@@ -110,7 +110,7 @@
 }
 
 @article{tfjs2019,
-  title={Tensorflow.js: Machine Learning for the Web and Beyond},
+  title={Tensor{F}low.js: Machine Learning for the Web and Beyond},
   author={Smilkov, Daniel and Thorat, Nikhil and Assogba, Yannick and Nicholson, Charles and Kreeger, Nick and Yu, Ping and Cai, Shanqing and Nielsen, Eric and Soegel, David and Bileschi, Stan and others},
   journal={Proceedings of Machine Learning and Systems},
   volume={1},
