### 2025-12-18 6ad975e
Update paper/paper.bib
