+++ paper/paper.bib
@@ -130,7 +130,7 @@
 
 @misc{NerdCorner2025,
   author    = {{Nerd Corner}},
-  title     = {Tensorflow.js vs {T}ensorflow ({P}ython)},
+  title     = {Tensor{F}low.js vs {T}ensor{F}low ({P}ython)},
   year      = {2025},
   month     = {Mar},
   howpublished = {\url{https://nerd-corner.com/tensorflow-js-vs-tensorflow-python/}}
