### 2025-12-18 f14690d
Update paper.bib (#255)
* Update paper.bib

* Update paper/paper.bib

* Update paper/paper.bib
