### 2025-12-20 b85dd32
Refine citations in paper.md for clarity and consistency. (#256)
