### 2025-12-20 a9d8f5d
Refine citations in paper.md for clarity and consistency.
