### 2025-12-22 e7969c2
update `cite` section
