### 2025-12-23 f069a90
add `MinMaxScaler` into preprocessing table
