### 2025-12-23 327d7ed
add support for `compose` models
