### 2025-12-29 df9e08e
removing the table and using the JOSS badge without the table
