### 2025-12-30 302ea8b
refactor to import from `util`
