### 2025-12-30 bc33222
`compose` chain implemented
