### 2025-12-30 881cbe6
Add support for serializing and deserializing sklearn's _RemainderColsList and slice objects
