+++ pymilo/transporters/general_data_structure_transporter.py
@@ -14,11 +14,29 @@ from .transporter import AbstractTransporter
 class GeneralDataStructureTransporter(AbstractTransporter):
     """Customized PyMilo Transporter developed to handle fields with general datastructures."""
 
+    def _is_remainder_cols_list(self, obj):
+        """
+        Check whether the given object is sklearn's internal _RemainderColsList.
+
+        :param obj: given object
+        :type obj: any
+        :return: bool
+        """
+        try:
+            return (
+                obj.__class__.__name__ == "_RemainderColsList" and
+                "_column_transformer" in obj.__class__.__module__
+            )
+        except Exception:
+            return False
+
     def serialize_tuple(self, tuple_field):
         """
         Check for non-serializable fields in tuple and serialize them.
 
             1. Serialize inner np.ndarray fields in tuple
+            2. Serialize inner slice fields in tuple
+            3. Convert sklearn's _RemainderColsList to list
 
         :param tuple_field: given tuple
         :type tuple_field: tuple
@@ -28,6 +46,15 @@ class GeneralDataStructureTransporter(AbstractTransporter):
         for item in tuple_field:
             if (isinstance(item, np.ndarray)):
                 new_tuple += (self.deep_serialize_ndarray(item),)
+            elif self._is_remainder_cols_list(item):
+                new_tuple += (list(item),)
+            elif isinstance(item, slice):
+                new_tuple += ({
+                    "pymilo-slice": True,
+                    "start": item.start,
+                    "stop": item.stop,
+                    "step": item.step,
+                },)
             else:
                 new_tuple += (item,)
         return {
@@ -51,6 +78,17 @@ class GeneralDataStructureTransporter(AbstractTransporter):
             # check inner field as a np.ndarray
             if isinstance(dictionary[key], np.ndarray):
                 dictionary[key] = self.deep_serialize_ndarray(dictionary[key])
+            # check inner field as sklearn's _RemainderColsList
+            if self._is_remainder_cols_list(dictionary[key]):
+                dictionary[key] = list(dictionary[key])
+            # check inner field as a slice
+            if isinstance(dictionary[key], slice):
+                dictionary[key] = {
+                    "pymilo-slice": True,
+                    "start": dictionary[key].start,
+                    "stop": dictionary[key].stop,
+                    "step": dictionary[key].step,
+                }
             # check inner field as np.int32
             if isinstance(key, np.int32):
                 new_value = {
@@ -142,6 +180,17 @@ class GeneralDataStructureTransporter(AbstractTransporter):
                         {"value": int(item), "np-type": "numpy.int64"})
                 elif isinstance(item, np.ndarray):
                     new_list.append(self.deep_serialize_ndarray(item))
+                elif isinstance(item, tuple):
+                    new_list.append(self.serialize_tuple(item))
+                elif self._is_remainder_cols_list(item):
+                    new_list.append(list(item))
+                elif isinstance(item, slice):
+                    new_list.append({
+                        "pymilo-slice": True,
+                        "start": item.start,
+                        "stop": item.stop,
+                        "step": item.step
+                    })
                 else:
                     new_list.append(item)
             data[key] = new_list
@@ -156,6 +205,14 @@ class GeneralDataStructureTransporter(AbstractTransporter):
 
         elif isinstance(data[key], dict):
             data[key] = self.serialize_dict(data[key])
+
+        elif isinstance(data[key], slice):
+            data[key] = {
+                "pymilo-slice": True,
+                "start": data[key].start,
+                "stop": data[key].stop,
+                "step": data[key].step
+            }
 
[…]
