### 2025-12-30 4b90eac
`TransformedTargetRegressor` testcases
