### 2025-12-30 d5e8b9a
add test runner for compose models
