### 2025-12-30 9112b26
Remove unused imports and simplify `column_transformer` test by eliminating pandas dependency in favor of numpy arrays
