+++ pymilo/chains/ensemble_chain.py
@@ -20,7 +20,6 @@ from ..transporters.treepredictor_transporter import TreePredictorTransporter
 from ..pymilo_param import SKLEARN_ENSEMBLE_TABLE
 from ..utils.util import check_str_in_iterable
 from .util import serialize_possible_ml_model, deserialize_possible_ml_model
-from .util import get_concrete_transporter
 
 ENSEMBLE_CHAIN = {
     "FeatureExtractorTransporter": FeatureExtractorTransporter(),
