### 2025-12-30 0259a7a
add composite models to supported models
