+++ SUPPORTED_MODELS.md
@@ -1,6 +1,6 @@
 # Supported Models
 
-**Last Update: 2025-2-15**
+**Last Update: 2025-12-30**
 
 
 <h2 id="scikit-learn">Scikit-Learn</h2> 
@@ -776,5 +776,25 @@
 		<td>7</td>
 		<td><b>TfidfVectorizer</b></td>
 		<td>>=1.3</td>
+	</tr>
+</table>
+
+<h3 id="scikit-learn-feature-extraction">Composite Modules</h3> 
+📚 <a href="https://scikit-learn.org/stable/api/sklearn.compose.html" target="_blank"><b>Models Document</b></a>
+<table>
+	<tr align="center">
+		<th>ID</th>
+		<th>Model Name</th>
+        <th>PyMilo Version</th>
+	</tr>
+	<tr align="center">
+		<td>1</td>
+		<td><b>ColumnTransformer</b></td>
+		<td>>=1.5</td>
+	</tr>
+	<tr align="center">
+		<td>2</td>
+		<td><b>TransformedTargetRegressor</b></td>
+		<td>>=1.5</td>
 	</tr>
 </table>
