### 2025-12-30 92db6d0
Refactor model directory reset logic in tests to improve file handling and avoid unnecessary directory removal
