### 2025-12-30 3781dd3
Add `JOSS` published paper for `PyMilo` (#261)
* Refine citations in paper.md for clarity and consistency.

* Refine citations in paper.md for clarity and consistency.

* add pdf file of joss paper

* add `PyMilo` paper bib

* update `cite` section

* Update CHANGELOG.md

* Add CITATION.cff file for PyMilo, including metadata, authors, and preferred citation details

* Add .gitignore file to exclude refs directory

* removing the table and using the JOSS badge without the table
