+++ tests/test_ensembles/test_ensembles.py
@@ -25,6 +25,7 @@ from voting.voting_regressor import voting_regressor
 from voting.voting_classifier import voting_classifier
 
 from pipeline import pipeline
+from tests.test_composes.transformed_target_regressor import complex_transformed_target_regressor
 
 from pymilo.pymilo_param import SKLEARN_ENSEMBLE_TABLE, NOT_SUPPORTED
 
@@ -44,7 +45,7 @@ ENSEMBLES = {
     "Forests": [random_forest_regressor, random_forest_classifier, isolation_forest, random_trees_embedding],
     "Stacking": [stacking_regressor, stacking_classifier],
     "Voting": [voting_regressor, voting_classifier],
-    "Pipeline": [pipeline],
+    "Pipeline": [pipeline, complex_transformed_target_regressor],
 }
 
 @pytest.fixture(scope="session", autouse=True)
