### 2025-12-30 ad99849
apply pep8 docstring feedback
