+++ pymilo/transporters/compose_transporter.py
@@ -20,8 +20,10 @@ class ComposeTransporter(AbstractTransporter):
 
     def is_compose_internal_model(self, internal_model):
         """
-        Check whether the given content is a nested estimator/module that should be (de)serialized
-        when found inside compose models (e.g., inside ColumnTransformer.transformers).
+        Check whether the given content is a nested estimator/module.
+
+        This is used to detect objects that should be (de)serialized when found inside
+        compose models (e.g., inside ColumnTransformer.transformers).
 
         :param internal_model: given object
         :type internal_model: any
@@ -128,6 +130,21 @@ class ComposeTransporter(AbstractTransporter):
         return obj
 
     def serialize(self, data, key, model_type):
+        """
+        Serialize Compose object.
+
+        Serialize the data[key] of the given model which type is model_type.
+        basically in order to fully serialize a model, we should traverse over all the keys of its data dictionary and
+        pass it through the chain of associated transporters to get fully serialized.
+
+        :param data: the internal data dictionary of the given model
+        :type data: dict
+        :param key: the special key of the data param, which we're going to serialize its value(data[key])
+        :type key: object
+        :param model_type: the model type of the ML model, which data dictionary is given as the data param
+        :type model_type: str
+        :return: pymilo serialized output of data[key]
+        """
         if isinstance(data[key], (dict, list, tuple)):
             return self._serialize_nested(data[key])
         if self.is_compose_internal_model(data[key]):
@@ -135,6 +152,22 @@ class ComposeTransporter(AbstractTransporter):
         return data[key]
 
     def deserialize(self, data, key, model_type):
+        """
+        Deserialize previously pymilo serialized compose object.
+
+        Deserialize the data[key] of the given model which type is model_type.
+        basically in order to fully deserialize a model, we should traverse over all the keys of its serialized data dictionary and
+        pass it through the chain of associated transporters to get fully deserialized.
+
+        :param data: the internal data dictionary of the associated json file of the ML model which is generated previously by
+        pymilo export.
+        :type data: dict
+        :param key: the special key of the data param, which we're going to deserialize its value(data[key])
+        :type key: object
+        :param model_type: the model type of the ML model, which internal serialized data dictionary is given as the data param
+        :type model_type: str
+        :return: pymilo deserialized output of data[key]
+        """
         if isinstance(data[key], (dict, list, tuple)):
             return self._deserialize_nested(data[key])
         if self.is_compose_internal_model(data[key]):
