+++ CHANGELOG.md
@@ -31,6 +31,62 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `get_transporter` in `ensemble_chain.py`
 - `deserialize_possible_ml_model` in `ensemble_chain.py`
 - `serialize_possible_ml_model` in `ensemble_chain.py`
+## [1.4] - 2025-12-01
+### Added
+- `get_allowed_models` function in `PymiloClient`
+- `get_allowance` function in `PymiloClient`
+- `revoke_access` function in `PymiloClient`
+- `grant_access` function in `PymiloClient`
+- `get_ml_models` function in `PymiloClient`
+- `deregister_ml_model` function in `PymiloClient`
+- `register_ml_model` function in `PymiloClient`
+- `deregister` function in `PymiloClient`
+- `register` function in `PymiloClient`
+- `REST_API_PREFIX ` function in `streaming.param.py`
+- `register_client` function in `RESTClientCommunicator`
+- `remove_client` function in `RESTClientCommunicator`
+- `register_model` function in `RESTClientCommunicator`
+- `remove_model` function in `RESTClientCommunicator`
+- `get_ml_models` function in `RESTClientCommunicator`
+- `grant_access` function in `RESTClientCommunicator`
+- `revoke_access` function in `RESTClientCommunicator`
+- `get_allowance` function in `RESTClientCommunicator`
+- `get_allowed_models` function in `RESTClientCommunicator`
+- `_validate_id` function in `PymiloServer`
+- `init_client` function in `PymiloServer`
+- `remove_client` function in `PymiloServer`
+- `grant_access` function in `PymiloServer`
+- `revoke_access` function in `PymiloServer`
+- `get_allowed_models` function in `PymiloServer`
+- `get_clients_allowance` function in `PymiloServer`
+- `get_clients` function in `PymiloServer`
+- `init_ml_model` function in `PymiloServer`
+- `set_ml_model` function in `PymiloServer`
+- `remove_ml_model` function in `PymiloServer`
+- `get_ml_models` function in `PymiloServer`
+### Changed
+- `is_callable_attribute` function in `PymiloServer`
+- `execute_model` function in `PymiloServer`
+- `update_model` function in `PymiloServer`
+- `export_model` function in `PymiloServer`
+- `__getattr__` in `PymiloClient`
+- `upload` function in `PymiloClient`
+- `download` function in `PymiloClient`
+- `encrypt_compress` function in `PymiloClient`
+- `ClientCommunicator` interface
+- `handle_message` function in `WebSocketServerCommunicator`
+- `_handle_download` function in `WebSocketServerCommunicator`
+- `setup_routes` function in `RESTServerCommunicator`
+- `__init__` function in `RESTClientCommunicator`
+- `download` function in `RESTClientCommunicator`
+- `upload` function in `RESTClientCommunicator`
+- `attribute_call` function in `RESTClientCommunicator`
+- `attribute_type` function in `RESTClientCommunicator`
+- `README.md` updated
+- `__init__` function in `PyMiloServer`
+- Test system modified
+- `Python 3.14` added to `test.yml`
+### Removed
 - Python 3.6 support
 ## [1.3] - 2025-02-26
 ### Added
