### 2026-01-01 0761fda
Trailing whitespaces + minor refactoring
