### 2026-01-19 1ca60b5
Refactor ComposeModelChain: remove serialize method and streamline deserialize logic
