### 2026-01-19 55e091b
Refactor ComposeTransporter: improve serialization and deserialization docstrings, streamline model type checks
