### 2026-01-19 362fddc
Update docstrings in util.py for clarity on return values of transporter functions
