### 2026-01-19 53575f7
Enhance docstrings in util.py to clarify return values for serialization and deserialization functions
