### 2026-01-19 682fd58
Enhance FunctionTransporter to support both FunctionType and BuiltinFunctionType for improved function handling
