### 2026-01-19 3e47bb4
Refactor PreprocessingTransporter to iterate over all transporters for serialization and deserialization, enhancing flexibility in data handling
