### 2026-01-19 bcd13c6
Refactor PreprocessingTransporter to support both old and new attribute names for BSpline, improving compatibility with different data formats
