### 2026-01-22 9d87053
Add sklearn compose models (#264)
* add `MinMaxScaler` into preprocessing table

* add support for `compose` models

* refactor to import from `util`

* `compose` chain implemented

* refactor and remove `get_transporter`

* `compose` transporter implemented

* Add support for serializing and deserializing sklearn's _RemainderColsList and slice objects

* add `get_transporter` into `util.py`

* `TransformedTargetRegressor` testcases

* `ColumnTransformer` testcases

* refactor import in `test_pymilo.py`

* add test runner for compose models

* add util for compose model run

* Refactor imports
[…]
