+++ pymilo/transporters/compose_transporter.py
@@ -1,0 +1,185 @@
+# -*- coding: utf-8 -*-
+"""PyMilo Compose transporter."""
+
+from ..utils.util import check_str_in_iterable
+from .transporter import AbstractTransporter
+from .general_data_structure_transporter import GeneralDataStructureTransporter
+from .preprocessing_transporter import PreprocessingTransporter
+from .feature_extraction_transporter import FeatureExtractorTransporter
+from ..chains.util import serialize_possible_ml_model, deserialize_possible_ml_model
+
+COMPOSE_CHAIN = {
+    "GeneralDataStructureTransporter": GeneralDataStructureTransporter(),
+    "PreprocessingTransporter": PreprocessingTransporter(),
+    "FeatureExtractorTransporter": FeatureExtractorTransporter(),
+}
+
+
+class ComposeTransporter(AbstractTransporter):
+    """Compose object dedicated Transporter."""
+
+    def is_compose_internal_model(self, internal_model):
+        """
+        Check whether the given content is a nested estimator/module.
+
+        This is used to detect objects that should be (de)serialized when found inside
+        compose models (e.g., inside ColumnTransformer.transformers).
+
+        :param internal_model: given object
+        :type internal_model: any
+        :return: bool
+        """
+        pt = COMPOSE_CHAIN["PreprocessingTransporter"]
+        fe = COMPOSE_CHAIN["FeatureExtractorTransporter"]
+        if isinstance(internal_model, dict):
+            return (
+                check_str_in_iterable("pymilo-inner-model-type", internal_model) or
+                pt.is_preprocessing_module(internal_model) or
+                fe.is_fe_module(internal_model)
+            )
+        return (
+            self._is_ml_model(internal_model) or
+            pt.is_preprocessing_module(internal_model) or
+            fe.is_fe_module(internal_model)
+        )
+
+    def _is_ml_model(self, obj):
+        """
+        Check if the object is an ML model that needs serialization.
+
+        :param obj: given object
+        :type obj: any
+        :return: bool
+        """
+        return hasattr(obj, 'fit') and (hasattr(obj, 'predict') or hasattr(obj, 'transform'))
+
+    def serialize_compose_internal_model(self, internal_model):
+        """
+        Serialize internal model of compose objects.
+
+        :param internal_model: given sklearn internal model
+        :type internal_model: sklearn model or function
+        :return: pymilo serialized internal_model
+        """
+        # Handle preprocessing modules
+        pt = COMPOSE_CHAIN["PreprocessingTransporter"]
+        if pt.is_preprocessing_module(internal_model):
+            return pt.serialize_pre_module(internal_model)
+
+        # Handle feature extraction modules
+        fe = COMPOSE_CHAIN["FeatureExtractorTransporter"]
+        if fe.is_fe_module(internal_model):
+            return fe.serialize_fe_module(internal_model)
+
+        # Handle ML models (including compose/ensemble/concrete estimators) using the project-wide schema.
+        if self._is_ml_model(internal_model):
+            has_ml_model, result = serialize_possible_ml_model(internal_model)
+            if has_ml_model:
+                return result
+        return internal_model
+
+    def deserialize_compose_internal_model(self, serialized_internal_model):
+        """
+        Deserialize internal model of compose objects.
+
+        :param serialized_internal_model: serialized internal model(by pymilo)
+        :type serialized_internal_model: dict
+        :return: retrieved associated sklearn internal model
+        """
+        # Preprocessing / feature extraction modules
+        pt = COMPOSE_CHAIN["PreprocessingTransporter"]
+        if pt.is_preprocessing_module(serialized_internal_model):
[…]
