+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -85,6 +85,7 @@ body:
       label: PyMilo version
       description: Which version of PyMilo are you using?
       options:
+        - PyMilo 1.5
         - PyMilo 1.4
         - PyMilo 1.3
         - PyMilo 1.2
