+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [1.4] - 2026-01-26
+## [1.5] - 2026-01-26
 ### Added
 - `_is_remainder_cols_list` function in GeneralDataStructureTransporter
 - `ComposeTransporter` Transporter
