### 2026-01-25 6697bb4
Version 1.5 (#265)
* `CHANGELOG.md` updated

* update version to `1.5`

* Update CHANGELOG.md
