+++ otherfiles/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "pymilo" %}
-{% set version = "1.4" %}
+{% set version = "1.5" %}
 
 package:
     name: {{ name|lower }}
